# docs/api-documentation-openapi.json
{
    "openapi": "3.0.1",
...
    "info": {
        "title": "mimoto",
        "description": "mimoto",
        "license": {
            "name": "Mosip",
            "url": "https://docs.mosip.io/platform/license"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://api.collab.mosip.net/v1/mimoto",
            "description": "For Swagger"
        }
    ],
    "paths": {
...
        "/credentialshare/download": {
            "post": {
                "operationId": "download",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CredentialDownloadResponseDTO"
                                },
                                "examples": {
                                    "success response": {
                                        "value": {
                                            "credential": {
                                                "biometrics": {
                                                    "face": "<face data>"
                                                },
                                                "province": [
                                                    {
                                                        "language": "eng",
                                                        "value": "Kenitra"
                                                    },
                                                    {
                                                        "language": "kan",
                                                        "value": "ಕೆನಿತ್ರಾ"
                                                    }
                                                ],
                                                "phone": "4567897654",
                                                "city": [
                                                    {
                                                        "language": "eng",
                                                        "value": "Kenitra"
                                                    },
                                                    {
                                                        "language": "kan",
                                                        "value": "ಕೆನಿತ್ರಾ"
                                                    }
                                                ],
                                                "postalCode": "14022",
                                                "fullName": [
                                                    {
                                                        "language": "eng",
                                                        "value": "kimgḻbb"
                                                    },
                                                    {
                                                        "language": "kan",
                                                        "value": "ಕಿಮ್ಗ್ೞ್ಬ್ಬ್"
                                                    }
                                                ],
                                                "region": [
                                                    {
                                                        "language": "eng",
                                                        "value": "Rabat Sale Kenitra"
                                                    },
                                                    {
                                                        "language": "kan",
                                                        "value": "ರಬತ್ ಸಲೆ ಕೆನಿತ್ರ"
                                                    }
                                                ]
                                            },
                                            "verifiableCredential": {
                                                "@context": [
                                                    "https://www.w3.org/2018/credentials/v1",
                                                    "http://localhost:8088/.well-known/mosip-context.json",
                                                    {
                                                        "sec": "https://w3id.org/security#"
                                                    }
                                                ],
                                                "credentialSubject": {
                                                    "biometrics": "<biometrics>",
                                                    "province": [
                                                        {
                                                            "language": "eng",
                                                            "value": "Kenitra"
                                                        },
                                                        {
                                                            "language": "kan",
                                                            "value": "ಕೆನಿತ್ರಾ"
                                                        }
                                                    ],
                                                    "city": [
                                                        {
                                                            "language": "eng",
                                                            "value": "Kenitra"
                                                        },
                                                        {
                                                            "language": "kan",
                                                            "value": "ಕೆನಿತ್ರಾ"
                                                        }
                                                    ],
                                                    "phone": "4567897654",
                                                    "postalCode": "14022",
                                                    "fullName": [
                                                        {
                                                            "language": "eng",
                                                            "value": "kimgḻbb"
                                                        },
                                                        {
                                                            "language": "kan",
                                                            "value": "ಕಿಮ್ಗ್ೞ್ಬ್ಬ್"
                                                        }
                                                    ],
                                                    "dateOfBirth": "1990/01/01",
                                                    "id": "http://credential.idrepo/credentials/2057135194",
                                                    "vcVer": "VC-V1",
                                                    "UIN": "2057135194",
                                                    "region": [
                                                        {
                                                            "language": "eng",
                                                            "value": "Rabat Sale Kenitra"
                                                        },
                                                        {
                                                            "language": "kan",
                                                            "value": "ರಬತ್ ಸಲೆ ಕೆನಿತ್ರ"
                                                        }
                                                    ],
                                                    "email": "santhosdss14@gmail.com"
                                                },
                                                "id": "http://credential.idrepo/credentials/a3afcf8c-e5a6-4782-9ce5-84740ec2f332",
                                                "issuanceDate": "2022-10-31T05:08:26.980Z",
                                                "issuer": "http://localhost:8088/.well-known/controller.json",
                                                "proof": {
                                                    "type": "RsaSignature2018",
                                                    "created": "2022-10-31T05:08:26Z",
                                                    "proofPurpose": "assertionMethod",
                                                    "verificationMethod": "http://localhost:8088/.well-known/public-key.json",
                                                    "jws": "<The jws signature>"
                                                },
                                                "type": [
                                                    "VerifiableCredential",
                                                    "MOSIPVerifiableCredential"
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "download VC",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "individualId": {
                                        "type": "string"
                                    },
                                    "requestId": {
                                        "type": "string"
                                    }
                                },
                                "x-examples": {
                                    "Example 1": {
                                        "individualId": "2057135194",
                                        "requestId": "2bc3bb7f-6156-46d1-ae03-bf7b76e0c257"
                                    }
                                }
                            },
                            "examples": {
                                "example-1": {
                                    "value": {
                                        "individualId": "2057135194",
                                        "requestId": "2bc3bb7f-6156-46d1-ae03-bf7b76e0c257"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Credential Download - OTP Flow"
                ],
                "summary": "Generate Credential -  Download Credential"
            }
        },
...
        "/credentials/download": {
            "post": {
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "grant_type": {
                                        "type": "string",
                                        "example": "authorization_code"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "GL9N1V-authorizationcodeyBkoaonjplms"
                                    },
                                    "redirect_uri": {
                                        "type": "string",
                                        "example": "https://injiweb.collab.mosip.net/redirect"
                                    },
                                    "code_verifier": {
                                        "type": "string",
                                        "example": "Sq6nZNToevcodeVerifierjXw70LCJ"
                                    },
                                    "issuer": {
                                        "type": "string",
                                        "example": "Mosip"
                                    },
                                    "credential": {
                                        "type": "string",
                                        "example": "MosipVerifiableCredential"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Trusted Issuers"
                ],
                "description": "This API is responsible for generating PDFs for the received VC content. It fetches display properties from the well-known configuration of the issuer and incorporates them into the predefined template of the PDF file.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                },
                                "examples": {
                                    "Failure response 1": {
                                        "value": {
                                            "id": null,
                                            "version": "v1",
                                            "str": null,
                                            "responsetime": "2022-10-31T05:07:34.789Z",
                                            "metadata": null,
                                            "response": null,
                                            "errors": [
                                                {
                                                    "errorCode": "RESIDENT-APP-036",
                                                    "errorMessage": "Invalid Credential Type Id"
                                                }
                                            ]
                                        }
                                    },
                                    "Failure response 2": {
                                        "value": {
                                            "id": null,
                                            "version": "v1",
                                            "str": null,
                                            "responsetime": "2022-10-31T05:07:34.789Z",
                                            "metadata": null,
                                            "response": null,
                                            "errors": [
                                                {
                                                    "errorCode": "RESIDENT-APP-026",
                                                    "errorMessage": "Api not accessible failure"
                                                }
                                            ]
                                        }
                                    },
                                    "Failure response 3": {
                                        "value": {
                                            "id": null,
                                            "version": "v1",
                                            "str": null,
                                            "responsetime": "2022-10-31T05:07:34.789Z",
                                            "metadata": null,
                                            "response": null,
                                            "errors": [
                                                {
                                                    "errorCode": "RESIDENT-APP-028",
                                                    "errorMessage": "error occured while signing pdf"
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "summary": "Download Credential As PDF",
                "x-stoplight": {
                    "id": "42bjevdfa0stx"
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CredentialDownloadResponseDTO": {
                "type": "object",
                "properties": {
                    "credential": {
                        "$ref": "#/components/schemas/JsonNode"
                    },
                    "verifiableCredential": {
                        "$ref": "#/components/schemas/JsonNode"
                    }
                }
            },
            "JsonNode": {
                "type": "object"
            },
...
        }
    },
...
}